feat(python): improve settings API and documentation
