+++ ports/python/lorentzian_classification/_types.py
@@ -5,7 +5,17 @@ from __future__ import annotations
 import re
 from dataclasses import dataclass
 
-from .core import RESULT_FIELDNAMES, Settings
+from .core import RESULT_FIELDNAMES
+from .settings import (
+    BOOLEAN_SETTING_FIELDS,
+    FEATURE_SLOTS,
+    FEATURE_TYPES,
+    FLOAT_SETTING_FIELDS,
+    INTEGER_SETTING_FIELDS,
+    PINE_SETTING_BOUNDS,
+    VALID_SOURCES,
+    Settings,
+)
 
 
 EXPORT_PACK_HEADER = "// Generated by lorentzian_classification export-pack"
@@ -81,54 +91,6 @@ PARITY_FIXTURES: tuple[tuple[str, bool], ...] = (
     ("pine_coinbase_btcusd_1d_limited_history.csv", False),
     ("pine_btcusd_h1_trimmed_limited_history.csv", False),
 )
-VALID_SOURCES = ("open", "high", "low", "close", "hl2", "hlc3", "ohlc4")
-FEATURE_SLOTS = ("f1", "f2", "f3", "f4", "f5")
-FEATURE_TYPES = ("RSI", "WT", "CCI", "ADX")
-BOOLEAN_SETTING_FIELDS = {
-    "include_full_history",
-    "use_volatility_filter",
-    "use_regime_filter",
-    "use_adx_filter",
-    "use_ema_filter",
-    "use_sma_filter",
-    "use_kernel_filter",
-    "use_kernel_smoothing",
-    "use_dynamic_exits",
-    "show_exits",
-    "use_worst_case",
-    "show_kernel_estimate",
-    "show_bar_colors",
-    "show_bar_predictions",
-    "use_atr_offset",
-    "use_confidence_gradient",
-    "show_trade_stats",
-}
-INTEGER_SETTING_FIELDS = {
-    "neighbors_count",
-    "max_bars_back",
-    "feature_count",
-    "color_compression",
-    "adx_threshold",
-    "ema_period",
-    "sma_period",
-    "kernel_h",
-    "kernel_x",
-    "kernel_lag",
-}
-FLOAT_SETTING_FIELDS = {"regime_threshold", "kernel_r", "bar_predictions_offset"}
-PINE_SETTING_BOUNDS: dict[str, tuple[float | int | None, float | int | None]] = {
-    "neighbors_count": (1, 100),
-    "feature_count": (2, 5),
-    "color_compression": (1, 10),
-    "regime_threshold": (-10, 10),
-    "adx_threshold": (0, 100),
-    "ema_period": (1, None),
-    "sma_period": (1, None),
-    "kernel_h": (3, None),
-    "bar_predictions_offset": (0, None),
-}
-
-
 @dataclass(frozen=True)
 class ParityFixture:
     name: str
